Rename: 커스텀 셀렉트, 로딩 스피너 common 폴더로 이동
