+++ components/Restaurant/RestaurantReviews/index.tsx
@@ -11,7 +11,7 @@ import { SortOptions } from '@/types/Review'
 import React, { useEffect, useMemo, useState } from 'react'
 import RestaurantReviewForm from './RestaurantReviewForm'
 import RestaurantReviewsList from './RestaurantReviewsList'
-import CustomSelectInput from '@/components/CustomSelectInput'
+import CustomSelectInput from '@/components/common/CustomSelectInput'
 import { cn } from '@/utils/cn'
 import { sortLabelToValueMap, sortValueToLabelMap } from '@/constants'
 import { useReviewStore } from '@/store/useReviewStore'
